Adds routes and cleans up server code.
