+++ server.js
@@ -4,8 +4,12 @@ const app = express();
 const hbs = require("express-handlebars");
 const bodyParser = require("body-parser");
 const fs = require("fs");
+const session = require('express-session');
+const redisStore = require('connect-redis')(session);
+const passport = require('passport');
 
 const Gallery = require("./models/gallerymodel.js");
+const gRouter = require('./routes/gallery.js');
 
 const PORT = process.env.PORT;
 if (!PORT) {
@@ -19,8 +23,15 @@ app.use(
   })
 );
 app.use(bodyParser.json());
+app.use(session({
+  store: new redisStore(),
+  secret: 'aaaaAAA',
+  resave: false,
+  saveUninitialized: false
+}));
+app.use(passport.initialize());
+app.use(passport.session());
 
-//HBS STUFF
 app.engine(
   "handlebars",
   hbs({
@@ -28,214 +39,6 @@ app.engine(
   })
 );
 app.set("view engine", "handlebars");
-
-app.get("/", (req, res) => {
-  return new Gallery()
-    .fetchAll()
-    .then(photo => {
-      console.log(photo.models);
-      let arr = [];
-      photo.models.forEach(i => {
-        arr.push(i.attributes);
-      });
-      return res.render("main", {
-        arr
-      });
-    })
-    .catch(err => {
-      console.log(err);
-      res.sendStatus(500);
-    });
-});
-
-app.get("/gallery", (req, res) => {
-  return new Gallery()
-    .fetchAll()
-    .then(photo => {
-      console.log(photo.models);
-      let arr = [];
-      photo.models.forEach(i => {
-        arr.push(i.attributes);
-      });
-      return res.render("main", {
-        arr
-      });
-    })
-    .catch(err => {
-      console.log(err);
-      res.sendStatus(500);
-    });
-});
-
-/*
- ***********************
- * ADD
- ***********************
- */
-app.get("/gallery/new", (req, res) => {
-  return new Gallery()
-    .fetchAll()
-    .then(gallerytable => {
-      return res.render("new", {
-        gallerytable
-      });
-    })
-    .catch(err => {
-      console.log(err);
-      res.sendStatus(500);
-    });
-});
-
-app.post("/gallery", (req, res) => {
-  const body = req.body;
-  return Gallery.forge({
-      author: body.author,
-      link: body.link,
-      description: body.description
-    })
-    .save(null, {
-      method: "insert"
-    })
-    .then(() => {
-      new Gallery({
-          link: body.link
-        })
-        .fetch()
-        .then(img => {
-          return res.redirect("/gallery");
-        });
-    });
-});
-
-app.get("/gallery/:id", (req, res) => {
-  let reqParams = req.params.id;
-  return new Gallery()
-    .where({
-      id: reqParams
-    })
-    .fetch()
-    .then(img => {
-      let photo = img.attributes;
-
-      return res.render("id", photo);
-    })
-    .catch(err => {
-      console.log(err);
-      res.sendStatus(500);
-    });
-});
-
-app.get("/gallery/:id/edit", (req, res) => {
-  let paramsId = req.params.id;
-  return Gallery.where({
-      id: paramsId
-    })
-    .fetch()
-    .then(img => {
-      console.log(img.attributes);
-      let id = img.attributes.id;
-      let galleryObj = img.attributes;
-      res.render("edit", galleryObj);
-    });
-});
-
-/*
- ***********************
- * EDIT
- ***********************
- */
-
-app.put("/gallery/:id", (req, res) => {
-  const body = req.body;
-  const paramsId = req.params.id;
-
-  Gallery.where({
-      id: paramsId
-    })
-    .fetch()
-    .then(img => {
-      new Gallery({
-          id: paramsId
-        })
-        .save({
-          link: body.link,
-          description: body.description,
-          author: body.author
-        }, {
-          patch: true
-        })
-        .then(() => {
-          return res.redirect("/gallery");
-        });
-    });
-});
-
-app.post("/gallery/:id", (req, res) => {
-  const body = req.body;
-  const paramsId = req.params.id;
-
-  Gallery.where({
-      id: paramsId
-    })
-    .fetch()
-    .then(img => {
-      new Gallery({
-          id: paramsId
-        })
-        .save({
-          link: body.link,
-          description: body.description,
-          author: body.author
-        }, {
-          patch: true
-        })
-        .then(() => {
-          return res.redirect("/gallery");
-        });
-    });
-});
-
-/*
- ***********************
- * DELETE
- ***********************
- */
-
-app.delete("/gallery/:id", (req, res) => {
-  const paramsId = req.params.id;
-
-  Gallery.where({
-      id: paramsId
-    })
-    .fetch()
-    .then(img => {
-      new Gallery({
-          id: paramsId
-        })
-        .destroy()
-        .then(() => {
-          return res.redirect("/gallery");
-        });
-    });
-});
-
-app.get("/gallery/:id/delete", (req, res) => {
-  const paramsId = req.params.id;
-
-  Gallery.where({
-      id: paramsId
-    })
-    .fetch()
-    .then(img => {
-      new Gallery({
-          id: paramsId
-        })
-        .destroy()
-        .then(() => {
-          return res.redirect("/gallery");
-        });
-    });
-});
 
 app.get("/css/styles.css", (req, res) => {
   fs.readFile("./public/css/styles.css", (err, data) => {
@@ -247,16 +50,8 @@ app.get("/css/styles.css", (req, res) => {
   });
 });
 
-app.get("spiration_light.png", (req, res) => {
-  fs.readFile("./public/spiration_light.png", (err, data) => {
-    if (err) {
-      console.log(err);
-    }
-    res.write(data.toString());
-    res.end();
-  });
-});
-
+app.use('/gallery', gRouter);
+app.use('/', gRouter);
 
 app.listen(PORT, () => {
   console.log(`Listening on port ${PORT}!`);
